+++ spec/lockfile_parser_spec.rb
@@ -80,15 +80,15 @@ describe "the bundler version retrieved from the lockfile" do
     end
 
     it "uses the default version when we get a pessimistic qualifier and is lower than the default version" do
-      subject.fetch_version('1.0.1', '~>').should == '1.0.21'
+      subject.fetch_version('1.1.1', '~>').should == EY::Serverside::LockfileParser::DEFAULT
     end
 
     it "uses the given version when we get a pessimistic qualifier that doesn't match the default version" do
-      subject.fetch_version('1.1.0', '~>').should == '1.1.0'
+      subject.fetch_version('1.0.0', '~>').should == '1.0.0'
     end
 
     it "uses the given version when it's geater of equal than the default version" do
-      subject.fetch_version('1.1.0', '>=').should == '1.1.0'
+      subject.fetch_version('1.2.0', '>=').should == '1.2.0'
     end
 
     it "uses the default version when the given version is lower" do